# internal/docs/swagger.yaml
definitions:
...
  entity.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  entity.LoginResponse:
    properties:
      access_token:
        type: string
      id:
        type: integer
      refresh_token:
        type: string
      role:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
  entity.NewAccessTokenResponse:
    properties:
      access_token:
        type: string
      id:
        type: integer
      refresh_token:
        type: string
      role:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
...
  entity.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  entity.RegisterResponse:
    properties:
      access_token:
        type: string
      id:
        type: integer
      refresh_token:
        type: string
      role:
        type: string
      status:
        type: boolean
      username:
        type: string
    type: object
...
  errors.Error:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /v1/auth/login:
    post:
      consumes:
      - application/json
      description: This API for login
      parameters:
      - description: Login Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Login
      tags:
      - auth
  /v1/auth/new-access/{refresh}:
    get:
      consumes:
      - application/json
      description: This API for getting a new access-token
      parameters:
      - description: Refresh Token
        in: path
        name: refresh
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.NewAccessTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Get New Access
      tags:
      - auth
  /v1/auth/register:
    post:
      consumes:
      - application/json
      description: This API for validating email as register
      parameters:
      - description: Register Model
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.Error'
      summary: Register
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"